requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
openai 
pandas
openpyxl
PyPDF2
pypdf